for the f sake work pleas

+++ get_next_line_utils.c
@@ -6,7 +6,7 @@
 /*   By: mda-cunh <mda-cunh@42paris.fr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/12 18:53:58 by mda-cunh          #+#    #+#             */
-/*   Updated: 2023/11/12 21:15:09 by mda-cunh         ###   ########.fr       */
+/*   Updated: 2023/11/15 14:32:12 by mda-cunh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -57,7 +57,7 @@ char	*ft_strchr(const char *s, int c)
 
 size_t	ft_strlen(const char *s)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (s[i])
